feedback and fix issue
