+++ @xen-orchestra/web/src/components/pool/network/PoolHostInternalNetworksTable.vue
@@ -1,11 +1,18 @@
 <template>
-  <div class="pool-networks-table">
+  <div class="pool-host-internal-networks-table">
     <UiTitle>
-      {{ t('networks') }}
+      {{ t('host-internal-networks') }}
       <template #actions>
-        <UiDropdownButton v-tooltip="t('coming-soon')" disabled>
+        <UiButton
+          v-tooltip="t('coming-soon')"
+          disabled
+          :left-icon="faPlus"
+          variant="secondary"
+          accent="brand"
+          size="medium"
+        >
           {{ t('new') }}
-        </UiDropdownButton>
+        </UiButton>
       </template>
     </UiTitle>
     <div class="container">
@@ -21,16 +28,6 @@
             size="medium"
           >
             {{ t('edit') }}
-          </UiButton>
-          <UiButton
-            v-tooltip="t('coming-soon')"
-            disabled
-            :left-icon="faCopy"
-            variant="tertiary"
-            accent="brand"
-            size="medium"
-          >
-            {{ t('copy-info-json') }}
           </UiButton>
           <UiButton
             v-tooltip="t('coming-soon')"
@@ -94,7 +91,6 @@
                 disabled
                 size="small"
               />
-              <VtsConnectionStatus v-else-if="column.id === 'status'" :status="column.value" />
               <div v-else v-tooltip="{ placement: 'bottom-end' }" class="text-ellipsis">
                 {{ column.value }}
               </div>
@@ -111,34 +107,30 @@
 </template>
 
 <script setup lang="ts">
-import { useNetworkStore } from '@/stores/xo-rest-api/network.store'
-import { usePifStore } from '@/stores/xo-rest-api/pif.store'
-import type { XoNetwork } from '@/types/xo/network.type'
-import VtsConnectionStatus from '@core/components/connection-status/VtsConnectionStatus.vue'
+import { useNetworkStore } from '@/stores/xo-rest-api/network.store.ts'
+import type { XoNetwork } from '@/types/xo/network.type.ts'
 import VtsDataTable from '@core/components/data-table/VtsDataTable.vue'
 import VtsIcon from '@core/components/icon/VtsIcon.vue'
 import VtsStateHero from '@core/components/state-hero/VtsStateHero.vue'
 import UiButton from '@core/components/ui/button/UiButton.vue'
 import UiButtonIcon from '@core/components/ui/button-icon/UiButtonIcon.vue'
 import UiCheckbox from '@core/components/ui/checkbox/UiCheckbox.vue'
-import UiDropdownButton from '@core/components/ui/dropdown-button/UiDropdownButton.vue'
 import UiQuerySearchBar from '@core/components/ui/query-search-bar/UiQuerySearchBar.vue'
 import UiTableActions from '@core/components/ui/table-actions/UiTableActions.vue'
 import UiTitle from '@core/components/ui/title/UiTitle.vue'
 import UiTopBottomTable from '@core/components/ui/top-bottom-table/UiTopBottomTable.vue'
-import { useRouteQuery } from '@core/composables/route-query.composable'
-import useMultiSelect from '@core/composables/table/multi-select.composable'
-import { useTable } from '@core/composables/table.composable'
-import { vTooltip } from '@core/directives/tooltip.directive'
+import { useRouteQuery } from '@core/composables/route-query.composable.ts'
+import useMultiSelect from '@core/composables/table/multi-select.composable.ts'
+import { useTable } from '@core/composables/table.composable.ts'
+import { vTooltip } from '@core/directives/tooltip.directive.ts'
 import type { IconDefinition } from '@fortawesome/fontawesome-common-types'
 import {
   faAlignLeft,
   faCaretDown,
-  faCopy,
   faEdit,
   faEllipsis,
   faHashtag,
-  faPowerOff,
+  faPlus,
   faTrash,
 } from '@fortawesome/free-solid-svg-icons'
 import { noop } from '@vueuse/shared'
@@ -149,15 +141,11 @@ const { networks } = defineProps<{
   networks: XoNetwork[]
 }>()
 
-const { t } = useI18n()
-
 const { isReady, hasError } = useNetworkStore().subscribe()
 
-const { records: pifs } = usePifStore().subscribe()
-
-const selectedNetworkId = useRouteQuery('id')
-
+const { t } = useI18n()
 const searchQuery = ref('')
+const selectedNetworkId = useRouteQuery('id')
 
 const filteredNetworks = computed(() => {
   const searchTerm = searchQuery.value.trim().toLocaleLowerCase()
@@ -179,32 +167,6 @@ const toggleSelect = () => {
   selected.value = selected.value.length === 0 ? networkIds.value : []
 }
 
-const getNetworkVlan = (network: XoNetwork) => {
-  const networkPIFs = pifs.value.filter(pif => network.PIFs.includes(pif.id))
-
-  if (networkPIFs.length > 0) {
-    return networkPIFs[0].vlan !== -1 ? networkPIFs[0].vlan.toString() : t('none')
-  }
-}
-
-const getNetworkStatus = (network: XoNetwork) => {
-  const networkPIFs = pifs.value.filter(pif => network.PIFs?.includes(pif.id))
-
-  if (networkPIFs.length === 0) {
-    return 'disconnected'
-  }
-
-  const isConnected = networkPIFs.map(pif => pif.attached && pif.carrier)
-  if (isConnected.every(Boolean)) {
-    return 'connected'
-  }
-
-  if (isConnected.some(Boolean)) {
-    return 'partially-connected'
-  }
-  return 'disconnected'
-}
-
 const getLockingMode = (lockingMode: string) => (lockingMode === 'disabled' ? t('disabled') : t('unlocked'))
 
 const { visibleColumns, rows } = useTable('networks', filteredNetworks, {
@@ -212,11 +174,7 @@ const { visibleColumns, rows } = useTable('networks', filteredNetworks, {
   columns: define => [
     define('checkbox', noop, { label: '', isHideable: false }),
     define('name_label', { label: t('name') }),
-    define('name_description', {
-      label: t('description'),
-    }),
-    define('status', record => getNetworkStatus(record), { label: t('pifs-status') }),
-    define('vlan', record => getNetworkVlan(record), { label: t('vlan') }),
+    define('name_description', { label: t('description') }),
     define('MTU', { label: t('mtu') }),
     define('default_locking_mode', record => getLockingMode(record.default_locking_mode), {
       label: t('default-locking-mode'),
@@ -225,27 +183,25 @@ const { visibleColumns, rows } = useTable('networks', filteredNetworks, {
   ],
 })
 
-type NetworkHeader = 'name_label' | 'name_description' | 'status' | 'vlan' | 'MTU' | 'default_locking_mode'
+type NetworkHeader = 'name_label' | 'name_description' | 'MTU' | 'default_locking_mode'
 
 const headerIcon: Record<NetworkHeader, IconDefinition> = {
   name_label: faAlignLeft,
   name_description: faAlignLeft,
-  status: faPowerOff,
-  vlan: faAlignLeft,
   MTU: faHashtag,
   default_locking_mode: faCaretDown,
 }
 </script>
 
 <style scoped lang="postcss">
-.pool-networks-table,
+.pool-host-internal-networks-table,
 .table-actions,
 .container {
   display: flex;
   flex-direction: column;
 }
 
-.pool-networks-table {
+.pool-host-internal-networks-table {
   gap: 2.4rem;
 
   .container,
